Separate server and client code

+++ server/src/main/kotlin/home/playground/models/HashTagStatistic.kt
@@ -6,6 +6,6 @@ import java.io.Serializable
  * Created by cuong on 11/15/15.
  */
 data class HashTagStatistic(
-    val count: Long = 0,
+    val count  : Int = 0,
     val hashTag: String = ""
 ) : Serializable
